wasmer_config::app::snapshot_trigger

Enum SnapshotTrigger

pub enum SnapshotTrigger {

    Idle,
    FirstListen,
    FirstEnviron,
    FirstStdin,
    FirstSigint,
    PeriodicInterval,
    Sigint,
    Sigalrm,
    Sigtstp,
    Sigstop,
    NonDeterministicCall,
    Bootstrap,
    Transaction,
    Explicit,

}

Variants

Idle

Triggered when all the threads in the process goes idle

FirstListen

Triggered when a listen syscall is invoked on a socket for the first time

FirstEnviron

Triggered on reading the environment variables for the first time

FirstStdin

Triggered when the process reads stdin for the first time

FirstSigint

Issued on the first interrupt signal (Ctrl + C) the process receives, after that normal CTRL-C will apply.

PeriodicInterval

Triggered periodically based on a interval (default 10 seconds) which can be specified using the snapshot-interval option

Sigint

Issued if the user sends an interrupt signal (Ctrl + C).

Sigalrm

Alarm clock signal (used for timers)

Sigtstp

The SIGTSTP signal is sent to a process by its controlling terminal to request it to stop temporarily. It is commonly initiated by the user pressing Ctrl-Z.

Sigstop

The SIGSTOP signal instructs the operating system to stop a process for later resumption.

NonDeterministicCall

When a non-determinstic call is made

Bootstrap

Bootstrapping process

Transaction

Transaction

Explicit

Explicitly requested by the guest module

Implementations

impl SnapshotTrigger

pub fn to_str(&self) -> &'static str