wasmer_config::app

Struct AppConfigV1

pub struct AppConfigV1 {

    pub name: Option<String>,
    pub app_id: Option<String>,
    pub owner: Option<String>,
    pub package: PackageSource,
    pub domains: Option<Vec<String>>,
    pub locality: Option<Locality>,
    pub env: IndexMap<String, String>,
    pub cli_args: Option<Vec<String>>,
    pub capabilities: Option<AppConfigCapabilityMapV1>,
    pub scheduled_tasks: Option<Vec<AppScheduledTask>>,
    pub volumes: Option<Vec<AppVolume>>,
    pub health_checks: Option<Vec<HealthCheckV1>>,
    pub debug: Option<bool>,
    pub scaling: Option<AppScalingConfigV1>,
    pub redirect: Option<Redirect>,
    pub jobs: Option<Vec<Job>>,
    pub extra: IndexMap<String, Value>,

}

User-facing app.yaml config file for apps.

NOTE: only used by the backend; Edge itself does not use this format and relies on the internal AppVersionV1Spec representation instead.

Fields
name: Option<String>

Name of the app.

app_id: Option<String>

App id assigned by the backend.

This will get populated once the app has been deployed.

This id is also used to map to the existing app during deployments.

owner: Option<String>

Owner of the app.

This is either a username or a namespace.

package: PackageSource

The package to execute.

domains: Option<Vec<String>>

Domains for the app.

This can include both provider-supplied alias domains and custom domains.

locality: Option<Locality>

Location-related configuration for the app.

env: IndexMap<String, String>

Environment variables.

cli_args: Option<Vec<String>>

Only applicable for runners that accept CLI arguments.

capabilities: Option<AppConfigCapabilityMapV1>scheduled_tasks: Option<Vec<AppScheduledTask>>volumes: Option<Vec<AppVolume>>health_checks: Option<Vec<HealthCheckV1>>debug: Option<bool>

Enable debug mode, which will show detailed error pages in the web gateway.

scaling: Option<AppScalingConfigV1>redirect: Option<Redirect>jobs: Option<Vec<Job>>extra: IndexMap<String, Value>

Capture extra fields for forwards compatibility.

Implementations

impl AppConfigV1

pub const KIND: &'static str = "wasmer.io/App.v0"

pub const CANONICAL_FILE_NAME: &'static str = "app.yaml"

pub fn to_yaml_value(self) -> Result<Value, Error>

pub fn to_yaml(self) -> Result<String, Error>

pub fn parse_yaml(value: &str) -> Result<Self, Error>