wasmer_config

Module cargo_annotations

Rust and cargo specific annotations used to interoperate with external tools.

Structs

CargoWasmerPackageAnnotation

The annotation used by cargo wapm when it parses the [package.metadata.wapm] table in your Cargo.toml.