Module wasmer_config::package

Wasmer package definitions.

Describes the contents of a wasmer.toml file.

Structs

CommandV1Deprecated

Describes a command for a wasmer module.

CommandV2

An executable command.

FileCommandAnnotations

Annotations on disk.

Manifest

The manifest represents the file used to describe a Wasm package.

ManifestBuilder

Builder for Manifest.

Module

A file which may be executed by a Command. Sometimes also referred to as an “atom”.

NamedPackageId

NamedPackageIdent

Parsed representation of a package identifier.

Package

Package definition for a Wasmer package.

PackageBuilder

Builder for Package.

PackageHash

Hash for a package.

PackageParseError

Error that occurs during package ident/source parsing.

WaiBindings

WitBindings

Enums

Abi

The ABI is a hint to WebAssembly runtimes about what additional imports to insert and how a module may be run.

Bindings

The interface exposed by a Module.

Command

CommandAnnotations

Annotations for a command.

FileKind

The different formats that FileCommandAnnotations can be saved in.

ImportsError

Errors that may occur when resolving Bindings imports.

ManifestBuilderError

Error type for ManifestBuilder

ManifestError

Errors that may occur while working with a Manifest.

ModuleReference

A reference to a module which may or may not come from another package.

PackageBuilderError

Error type for PackageBuilder

PackageId

PackageIdent

PackageSource

Source location of a package.

Tag

ValidationError

Errors that may be returned by Manifest::validate().

Statics

MANIFEST_FILE_NAME

The default name for the manifest file.