Enum wasmer_config::package::ModuleReference

pub enum ModuleReference {
    CurrentPackage {
        module: String,
    },
    Dependency {
        dependency: String,
        module: String,
    },
}

A reference to a module which may or may not come from another package.

Serialization

A ModuleReference is serialized via its String representation.

Variants

CurrentPackage

Fields

module: String

The name of the module.

A module in the current package.

Dependency

Fields

dependency: String

The name of the dependency the module comes from.

module: String

The name of the module.

A module that will be provided by a dependency, in dependency:module form.