wasmer_config::package::named_package_ident

Struct NamedPackageIdent

pub struct NamedPackageIdent {
    pub registry: Option<String>,
    pub namespace: Option<String>,
    pub name: String,
    pub tag: Option<Tag>,
}

Parsed representation of a package identifier.

Format: https?://<domain>/namespace/name@version where the registry, namespace, and version components are optional.

Fields
registry: Option<String>namespace: Option<String>name: Stringtag: Option<Tag>
Implementations

impl NamedPackageIdent

pub fn try_from_full_name_and_version( full_name: &str, version: &str, ) -> Result<Self, PackageParseError>

pub fn tag_str(&self) -> Option<String>

pub fn full_name(&self) -> String

Namespaced name.

Eg: “namespace/name”

pub fn version_opt(&self) -> Option<&VersionReq>

pub fn version_or_default(&self) -> VersionReq

pub fn registry_url(&self) -> Result<Option<Url>, PackageParseError>

pub fn build_identifier(&self) -> String

Build the ident for a package.

Format: NAMESPACE/NAME@tag where the namespace and tag components are optional.

pub fn build(&self) -> String

pub fn matches_id(&self, id: &NamedPackageId) -> bool

Returns true if this ident matches the given package id.

Semver constraints are matched against the package id’s version.