wasmer_config::package
Struct CommandV2

pub struct CommandV2 {
    pub name: String,
    pub module: ModuleReference,
    pub runner: String,
    pub annotations: Option<CommandAnnotations>,
}

An executable command.

Fields
name: String

The name of the command.

module: ModuleReference

The module containing this commandโs executable.

runner: String

The runner to use when running this command.

This may be a URL, or the well-known runners wasi or wcgi

annotations: Option<CommandAnnotations>

Extra annotations that will be consumed by the runner.

Implementations

impl CommandV2

pub fn get_annotations(&self, basepath: &Path) -> Result<Option<Value>, String>

Get annotations, automatically loading them from a file relative to the wasmer.tomlโs directory, if necessary.