wasmer_config::package

Struct ManifestBuilder

pub struct ManifestBuilder {
    package: Option<Option<Package>>,
    dependencies: Option<IndexMap<String, VersionReq>>,
    fs: Option<IndexMap<String, PathBuf>>,
    modules: Option<Vec<Module>>,
    commands: Option<Vec<Command>>,
}

Builder for Manifest.

Fields
package: Option<Option<Package>>

Metadata about the package itself.

dependencies: Option<IndexMap<String, VersionReq>>

The packageโs dependencies.

fs: Option<IndexMap<String, PathBuf>>

The mappings used when making bundled assets available to WebAssembly instances, in the form guest -> host.

modules: Option<Vec<Module>>

WebAssembly modules to be published.

commands: Option<Vec<Command>>

Commands the package makes available to users.

Implementations

impl ManifestBuilder

pub fn package(&mut self, value: Option<Package>) -> &mut Self

Metadata about the package itself.

pub fn dependencies(&mut self, value: IndexMap<String, VersionReq>) -> &mut Self

The packageโs dependencies.

pub fn fs(&mut self, value: IndexMap<String, PathBuf>) -> &mut Self

The mappings used when making bundled assets available to WebAssembly instances, in the form guest -> host.

pub fn modules(&mut self, value: Vec<Module>) -> &mut Self

WebAssembly modules to be published.

pub fn commands(&mut self, value: Vec<Command>) -> &mut Self

Commands the package makes available to users.

pub fn build(&self) -> Result<Manifest, ManifestBuilderError>

Builds a new Manifest.

Errors

If a required field has not been initialized.

fn create_empty() -> Self

Create an empty builder, with all fields set to None or PhantomData.

impl ManifestBuilder

pub fn new(package: Package) -> Self

pub fn map_fs( &mut self, guest: impl Into<String>, host: impl Into<PathBuf>, ) -> &mut Self

Include a directory on the host in the package and make it available to a WebAssembly guest at the guest path.

pub fn with_dependency( &mut self, name: impl Into<String>, version: VersionReq, ) -> &mut Self

Add a dependency to the Manifest.

pub fn with_module(&mut self, module: Module) -> &mut Self

Add a Module to the Manifest.

pub fn with_command(&mut self, command: Command) -> &mut Self

Add a Command to the Manifest.