wasmer_config::package
Struct Package

#[non_exhaustive]
pub struct Package {

    pub name: Option<String>,
    pub version: Option<Version>,
    pub description: Option<String>,
    pub license: Option<String>,
    pub license_file: Option<PathBuf>,
    pub readme: Option<PathBuf>,
    pub repository: Option<String>,
    pub homepage: Option<String>,
    pub wasmer_extra_flags: Option<String>,
    pub disable_command_rename: bool,
    pub rename_commands_to_raw_command_name: bool,
    pub entrypoint: Option<String>,
    pub private: bool,

}

Package definition for a Wasmer package.

Usually stored in a wasmer.toml file.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Option<String>

The package’s name in the form namespace/name.

version: Option<Version>

The package’s version number.

description: Option<String>

A brief description of the package.

license: Option<String>

A SPDX license specifier for this package.

license_file: Option<PathBuf>

The location of the license file, useful for non-standard licenses

readme: Option<PathBuf>

The package’s README file.

repository: Option<String>

A URL pointing to the package’s source code.

homepage: Option<String>

The website used as the package’s homepage.

wasmer_extra_flags: Option<String>
👎Deprecated since 0.9.2: Use runner-specific command attributes instead
disable_command_rename: bool
👎Deprecated since 0.9.2: Does nothing. Prefer a runner-specific command attribute instead
rename_commands_to_raw_command_name: bool
👎Deprecated since 0.9.2: Does nothing. Prefer a runner-specific command attribute instead

Unlike, disable-command-rename which prevents wasmer run <Module name>, this flag enables the command rename of wasmer run <COMMAND_NAME> into just <COMMAND_NAME>. This is useful for programs that need to inspect their argv[0] names and when the command name matches their executable name.

entrypoint: Option<String>

The name of the command that should be used by wasmer run by default.

private: bool

Mark this as a private package

Implementations

impl Package

pub fn new_empty() -> Self

pub fn builder( name: impl Into<String>, version: Version, description: impl Into<String>, ) -> PackageBuilder

Create a PackageBuilder populated with all mandatory fields.